+++ pclient/libs/file.h
@@ -7,7 +7,7 @@ void openFile(const char *filename, int numChannels, bool* sampleChannels);
 
 void closeFile();
 
-void writeToFile(int adcValue);
+void writeToFile(float adcValue);
 
 void writeNewLine();
